% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_init.R
\name{build_location_manager}
\alias{build_location_manager}
\title{Build Location Manager from Data}
\usage{
build_location_manager(location_data)
}
\arguments{
\item{location_data}{A list containing the serialized location data}
}
\value{
A fully initialized Location.Manager object
}
\description{
Internal function to construct a Location.Manager instance from
serialized data structures. This is called during package load.
}
\keyword{internal}
