% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_init.R
\name{extract_location_data}
\alias{extract_location_data}
\title{Extract Location Manager Data}
\usage{
extract_location_data(manager)
}
\arguments{
\item{manager}{A Location.Manager object}
}
\value{
A list containing all location data in simple structures
}
\description{
Internal function to extract all data from a Location.Manager instance
into serializable data structures (data.frames and lists).
}
\keyword{internal}
