% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.all.for.type}
\alias{get.all.for.type}
\title{get.all.for.type}
\usage{
get.all.for.type(type)
}
\arguments{
\item{type}{A string type name}
}
\value{
if 'type' is a registered type, it will return a vector of all location codes for that type.  If 'type' is not registered, it will return NA
}
\description{
Returns a vector of all valid location codes of a certain type
}
