% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.cbsa.for.msa.name}
\alias{get.cbsa.for.msa.name}
\title{get.cbsa.for.msa.name}
\usage{
get.cbsa.for.msa.name(names)
}
\arguments{
\item{names}{A vector of msa names to convert into cbsa location codes}
}
\value{
A named vector of string location codes, with NAs where no result was found
}
\description{
We are reading data in from .csv files, and these files do not have proper cbsa codes; instead the are indexed by cbsa name.  We need
            to take a vector of msa names and return a named vector of cbsa location codes.
}
