% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.containing.locations}
\alias{get.containing.locations}
\title{get.containing.locations}
\usage{
get.containing.locations(
  locations,
  super.type,
  return.list = F,
  throw.error.if.unregistered.type = T,
  include.partial = F
)
}
\arguments{
\item{locations}{A character vector of location codes}

\item{super.type}{The type (geographic resolution) of locations requested for the super-locations}

\item{return.list}{A single logical value indicating whether the return value should be a list with one element for each location, or whether all super-locations should be 'unlisted' into a vector}

\item{throw.error.if.unregistered.type}{A single logical value indicating whether the function should throw an error if super.type has not been registered as a location type}

\item{include.partial}{A single logical value indicating whether locations that only partially contain the input should also be returned. The default is FALSE to preserve complete-containment semantics.}
}
\value{
If return.list==T, a list with length(locations) and names=locations. Each element is itself a character vector with zero or more locations corresponding to super-locations. If return.list=F, returns a character vector (arbitrary length) containing all super-locations that contain ANY of the given locations. When include.partial is TRUE, directly registered partial overlaps are included.
}
\description{
Get Locations that Enclose a Location
}
