% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.location.code.if.unique}
\alias{get.location.code.if.unique}
\title{get.location.code.if.unique}
\usage{
get.location.code.if.unique(location.names, types, search.aliases = TRUE)
}
\arguments{
\item{location.names}{A list of names to get the location code for}

\item{types}{A corresponding list of types (or one type to be applied to each)}

\item{search.aliases}{A boolean flag indicating whether we want to check the aliases if the name check fails. Default = TRUE}
}
\value{
A named vector of location.codes, with names corresponding to the location names.  If there are multiple results or no results,
       the result in the vector will be NA
}
\description{
Get the location code for a name and a type.  The difference between this function and get.location.code
            is that get.location.code returns a list with multiple entries, and this function returns a named vector
            of results only if the location name matches one location uniquely.  This function uses get.location.code
            and converts the output.
}
