% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.location.code}
\alias{get.location.code}
\title{get.location.code}
\usage{
get.location.code(location.names, types, search.aliases = TRUE)
}
\arguments{
\item{location.names}{A list of names to get the location code for}

\item{types}{A corresponding list of types}

\item{search.aliases}{A boolean flag indicating whether we want to check the aliases if the name check fails. Default = TRUE}
}
\value{
A list of location.names, with values corresponding to the identified location code.  Multiple results per location.names is possible. If if the location.name is not registered or NA, NA is returned
}
\description{
Get the location code for a name and a type
}
