% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.location.coords}
\alias{get.location.coords}
\title{get.location.coords}
\usage{
get.location.coords(locations)
}
\arguments{
\item{locations}{A character vector of location codes}
}
\value{
A character vector of string location coordinates, separated by commas (latitude first, longitude second), with length(locations) and names=locations. If location codes are not registered (or if they were NA), the corresponding returned name is NA
}
\description{
Get the latitude and longitude of a location, provided it is known
}
