% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.location.name.alias}
\alias{get.location.name.alias}
\title{get.location.name.alias}
\usage{
get.location.name.alias(
  locations,
  alias.name,
  throw.error.if.unregistered.alias = T
)
}
\arguments{
\item{locations}{A character vector of location codes}

\item{alias.name}{A single character value representing a previously registered name alias}

\item{throw.error.if.unregistered.alias}{A single logical value indicating whether the function should throw an error if no alias with name alias.name has been registered}
}
\value{
A character vector of aliases, with length(locations) and names=locations. If location codes are not registered (or if they were NA), or if no value for the alias.name has been registered to a location, the corresponding returned alias is NA
}
\description{
Get an Name alias Associated with a Location
}
