% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.location.name}
\alias{get.location.name}
\title{get.location.name}
\usage{
get.location.name(locations)
}
\arguments{
\item{locations}{A character vector of location codes}
}
\value{
A character vector of location names, with length(locations) and names=locations. If location codes are not registered (or if they were NA), the corresponding returned name is NA
}
\description{
Get the Name of a Location
}
