% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.location.polygon}
\alias{get.location.polygon}
\title{get.location.polygon}
\usage{
get.location.polygon(location)
}
\arguments{
\item{locations}{A single location code}
}
\value{
A data frame containing the polygon data from the location code, or NA if not found
}
\description{
Get the polygon dataframe of a location, provided it is known
}
