% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.location.type}
\alias{get.location.type}
\title{get.location.type}
\usage{
get.location.type(locations)
}
\arguments{
\item{locations}{A character vector of location codes}
}
\value{
A character vector of location types, with length(locations) and names=locations. If location codes are not registered (or if they were NA), the corresponding returned type is NA
}
\description{
Get the Type (Geographic Resolution) of a Location
}
