% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.location.types}
\alias{get.location.types}
\title{get.location.types}
\usage{
get.location.types()
}
\value{
A list the same length as the registered number of types, with each list item being the name, the prefix and the longform prefix name.  If there are no registered types it will return an empty list
}
\description{
Get a list vector of registered location types, prefixes and longform prefix names.
}
