% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.overlapping.locations}
\alias{get.overlapping.locations}
\title{get.overlapping.locations}
\usage{
get.overlapping.locations(
  locations,
  type,
  return.list = F,
  throw.error.if.unregistered.type = T
)
}
\arguments{
\item{locations}{A character vector of location codes}

\item{type}{The type (geographic resolution) of locations requested for the sub-locations}

\item{return.list}{A single logical value indicating whether the return value should be a list with one element for each location, or whether all sub-locations should be 'unlisted' into a vector}

\item{throw.error.if.unregistered.type}{A single logical value indicating whether the function should throw an error if sub.type has not been registered as a location type}
}
\value{
If return.list==T, a list with length(locations) and names=locations. Each element is itself a character vector with zero or more locations corresponding to sub-locations. If return.list=F, returns a character vector (arbitrary length) containing all sub-locations that fall within ANY of the given locations
}
\description{
Get all locations that overlap a Location
}
