% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.polygons.for.type}
\alias{get.polygons.for.type}
\title{get.polygons.for.type}
\usage{
get.polygons.for.type(type)
}
\arguments{
\item{type}{A single location type}
}
\value{
A data.frame containing all the polygon data for a given type, NA if it doesn't exist
}
\description{
Return the polygon data for a given location type
}
