% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.prefix.for.type}
\alias{get.prefix.for.type}
\title{get.prefix.for.type}
\usage{
get.prefix.for.type(location.types)
}
\arguments{
\item{location.types}{A character vector of location types}
}
\value{
A character vector of prefixes, with length(location.types) and names=prefixes. If the types are not registered (or if they were NA), the corresponding returned type is NA
}
\description{
Get the prefix for a given type
}
