% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.sub.locations}
\alias{get.sub.locations}
\title{get.sub.locations}
\usage{
get.sub.locations(
  locations,
  sub.type,
  limit.to.completely.enclosing,
  return.list = F,
  throw.error.if.unregistered.type = T
)
}
\arguments{
\item{locations}{A character vector of location codes}

\item{sub.type}{The type (geographic resolution) of locations requested for the sub-locations}

\item{limit.to.completely.enclosing}{A single logical value indicating whether ONLY sub-locations that fall COMPLETELY within the given locations should be returned}

\item{return.list}{A single logical value indicating whether the return value should be a list with one element for each location, or whether all sub-locations should be 'unlisted' into a vector}

\item{throw.error.if.unregistered.type}{A single logical value indicating whether the function should throw an error if sub.type has not been registered as a location type}
}
\value{
If return.list==T, a list with length(locations) and names=locations. Each element is itself a character vector with zero or more locations corresponding to sub-locations. If return.list=F, returns a character vector (arbitrary length) containing all sub-locations that fall within ANY of the given locations
}
\description{
Get Locations that Fall Within a Location
}
