% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{LOCATION.MANAGER}
\alias{LOCATION.MANAGER}
\title{Package Lifecycle Hooks}
\format{
An object of class \code{environment} of length 36.
}
\usage{
LOCATION.MANAGER
}
\description{
This file (zzz.R) contains package lifecycle functions following R convention.
Functions here run during package load/unload/attach events.
}
\keyword{internal}
