% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_plot.R
\name{location.plot}
\alias{location.plot}
\title{location.plot}
\usage{
location.plot(
  data,
  color,
  fill,
  size = NA,
  title = NA,
  bb = c(left = -125, bottom = 24, right = -66, top = 50),
  bb.edge = 0.1,
  size.range = c(1, 5),
  color.range = c("blue", "red"),
  pch = 19,
  size.label = "",
  color.label = "",
  alpha = 1,
  map_water_color = "#C0C0C0",
  stadia_api_key = Sys.getenv("STADIA_MAPS_API_KEY")
)
}
\arguments{
\item{data}{A data frame with a column labeled 'location', containing a location code, and columns to be mapped
by color or size.}

\item{color}{The name of a column in the data frame to be used in the aesthetic as the color of the location.
If used with polygons, this is the outline color.}

\item{fill}{The name of a column in the data frame to be used as the fill color for points or polygons}

\item{size}{If left blank, the data will be plotted as polygons, and those locations without polygons will be skipped.
If filled in, it represents the name of a column in the data frame to be used to scale the size of the point,
and only those locations with latitude and longitude values will be plotted.}

\item{title}{Defaults to blank; a character string representing the title on the graph.}

\item{bb}{The bounding box value; The default is the bounding box of the contiguous US.  Otherwise, can take 
one of two values: either the character string "AUTO" which will use the dimensions of the locations to 
automatically bound the graph (using the bb.edge parameter, see below, as an edge buffer), or a named 
character vector of the format c(left=-125,bottom=24,right=-66, top=50), outlining the dimensions of 
the bounding box in latitude and longitude.  Defaults to NULL.}

\item{bb.edge}{If param 'bb' is 'AUTO', this will be used to create a buffer space around the outside of the 
automatically generated bounding box, as a proportion of the size of the plotted locations.  Defaults
to 0.1, or 10% of the bounding box.}

\item{size.range}{If 'size' is used, this will scale the values in the size column between the two size values 
specified.  Defaults to 1-5}

\item{color.range}{The color ranges to span for the 'color' and 'fill' parameters. Defaults to c('blue','red'),
so lower values will be more blue, and higher values will be more red.}

\item{pch}{The shape of the point when plotting by points; see https://sape.inf.usi.ch/quick-reference/ggplot2/shape
for a detailed list. Defaults to 19, a circle.}

\item{size.label}{A label to give to the size in the legend.  Defaults to blank.}

\item{color.label}{A label to give to the color/fill in the legend.  Defaults to blank.}

\item{alpha}{An alpha blending/transparency value for the 'color'/'fill' parameter. Defaults to 1, or no transparency.}

\item{map_water_color}{A hex color value for the color of any water on the map. Defaults to a light gray.}

\item{stadia_api_key}{The api key to access the stamen/stadia toner map tiles.  Defaults to looking for the
key in the STADIA_MAPS_API_KEY environment variable}
}
\value{
Returns the ggplot2 object/plot
}
\description{
Create a plot of various points in the US from a data frame
}
