% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{location.type.comprises}
\alias{location.type.comprises}
\title{location.type.comprises}
\usage{
location.type.comprises(super.location.type, sub.location.type)
}
\arguments{
\item{super.location.type}{The location type into which we are attempting to arrange another sub-type}

\item{sub.location.type}{The location type we are trying to fit inside the super-type}
}
\value{
A Boolean TRUE or FALSE, representing whether members of the sub.location.type fit perfectly in the super.location.type
}
\description{
This function allows us to understand the relationship between two different location types.  Some location types contain other
            location types completely (as states contain counties, for instance), while some are not comparable or disjoint in other ways (states and CBSAs).
            This function takes two location types (super.location.type and sub.location.type), and returns TRUE or FALSE depending on whether the types overlap 
            perfectly or not.
}
