% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{register.code.aliases}
\alias{register.code.aliases}
\title{register.code.aliases}
\usage{
register.code.aliases(location = NA, location.aliases = NA)
}
\arguments{
\item{location}{A single, previously registered location code or a registered location code alias.}

\item{location.aliases}{A character vector of location code aliases for this location name.  Code aliases are unique by type}
}
\description{
Register location code aliases for specific location
}
