% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{register.lat.and.long}
\alias{register.lat.and.long}
\title{register.lat.and.long}
\usage{
register.lat.and.long(location = NA, lat = NA, long = NA)
}
\arguments{
\item{location}{A previously registered location code, or a registered location code alias.}

\item{lat}{A single numeric value between -90 and 90}

\item{long}{A single numeric value between -180 and 180}
}
\description{
Register a valid latitude and longitude for a location code
}
