% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{register.locations}
\alias{register.locations}
\title{register.locations}
\usage{
register.locations(type, locations, location.names)
}
\arguments{
\item{type}{The geographic resolution at which to register the locations. Can be be either a single character value, or a vector of the same length as locations}

\item{locations}{A character vector of location codes}

\item{location.names}{A character vector of the same length as locations with corresponding names}
}
\description{
Register information about locations.
}
\details{
There is no error checking here; we assume that multiple locations with the same name are possible at different resolutions/types.
}
