% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{register.name.aliases}
\alias{register.name.aliases}
\title{register.name.aliases}
\usage{
register.name.aliases(
  location = NA,
  location.aliases = NA,
  location.aliases.names = NA
)
}
\arguments{
\item{location}{A single, previously registered location code or a registered location code alias.}

\item{location.aliases}{A character vector of aliases for this location name}

\item{location.alias.names}{A character vector names for the particular alias 'short','no spaces', 'full', etc}
}
\description{
Register name aliases for specific location
}
\details{
There is no error checking here; we assume that multiple locations with the same name are possible at different resolutions/types.
}
