% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{register.relationship.between.types}
\alias{register.relationship.between.types}
\title{register.relationships.between.types}
\usage{
register.relationship.between.types(super.type, sub.type, value)
}
\arguments{
\item{super.type}{A character vector of super types}

\item{sub.type}{A character vector of sub types of the same length}

\item{value}{A vector of boolean values indicating the whether the sub.type is contained completely in the super type, or not}
}
\description{
Register a relationship between location types; by default the types are independent; setting
            the value to TRUE indicates that the super.type contains the sub.type completely
}
