% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{register.sub.and.super.locations}
\alias{register.sub.and.super.locations}
\title{register.sub.and.super.locations}
\usage{
register.sub.and.super.locations(
  sub.locations,
  super.locations,
  super.completely.encloses.sub
)
}
\arguments{
\item{sub.locations}{A character vector of locations codes/location code aliases}

\item{super.locations}{A character vector of location codes/location code aliases of the same length as sub.locations, with corresponding super.locations}

\item{super.completely.encloses.sub}{Either a single logical value or a vector the same length as sub.locations and super.locations, indicating whether the location is completely enclosed}
}
\description{
Register hierarchical sub-super relationships
}
\details{
Where super.completely.encloses.sub==T, the function will automatically recognize that locations completely enclosed within the given sub.locations are also completely enclosed within the corresponding super.locations, and, conversely that locations which completely enclose the given super.locations also completely enclose the corresponding sub.locations
}
