% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{register.types}
\alias{register.types}
\title{register.types}
\usage{
register.types(type, prefix, prefix.longform)
}
\arguments{
\item{type}{A character vector representing types to be added}

\item{prefix}{A character vector of unique prefixes for the location codes for types of this kind}

\item{prefix.longform}{A character vector of longform names for that particular unique prefixes}
}
\description{
Register location type, prefix, and prefix.longform
}
\details{
The prefix is restricted to letters, numbers, period and '-'.
}
