% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{sanitize}
\alias{sanitize}
\title{sanitize}
\usage{
sanitize(codes)
}
\arguments{
\item{codes}{A character vector of location codes, or acceptable location codes}
}
\value{
A named character vector of location codes
}
\description{
There are values that we will accept as location codes (eg. 'c.12580' will work for Baltimore, but Baltimore's actual code is 'C.12580').
            This function converts codes we will accept into codes that are correct.  The function should fail on unrecognized location code.
}
