% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{get.code.by.alias}
\alias{get.code.by.alias}
\title{get.code.by.alias}
\usage{
get.code.by.alias(locations, types)
}
\arguments{
\item{locations}{A character vector of location code aliases}

\item{types}{A single type or a vector of types the same length as locations}
}
\value{
A character vector of true location codes, or NA
}
\description{
Get a location code by its code alias and type
}
