% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_location_manager.R
\name{get.location.types}
\alias{get.location.types}
\title{get.location.types}
\usage{
get.location.types(simple = TRUE)
}
\arguments{
\item{simple}{A boolean; TRUE if asking for simple list of types, FALSE if you want a list of types and prefixes.  Defaults to TRUE}
}
\value{
If simple=TRUE, a simple vector of named types; else a list the same length as the registered number of types, with each list item being the name, the prefix and the longform prefix name.  If there are no registered types it will return an empty list
}
\description{
Get a vector of type names, or a list of registered location types, prefixes and longform prefix names.
}
