% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_api.R
\name{is.location.valid}
\alias{is.location.valid}
\title{is.location.valid}
\usage{
is.location.valid(locations, suggest.options = F)
}
\arguments{
\item{locations}{A character vector representing potential locations}

\item{suggest.options}{A boolean indicating whether to check the aliases for potential matches}
}
\value{
A vector of boolean values whether the passed-in value is a location codes.  If false, this function will display a list of possibilities.
}
\description{
Check to see if the passed-in value matches any location code or alias
}
