% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCATIONS_plot.R
\name{location.plot}
\alias{location.plot}
\title{location.plot}
\usage{
location.plot(
  data,
  mapping,
  title = NA,
  size.range = c(1, 5),
  color.range = c("blue", "red"),
  pch = 19,
  size.label = "",
  color.label = "",
  alpha = 1
)
}
\arguments{
\item{data}{A data frame with a column labeled 'location', containing a location code}

\item{mapping}{Aesthetics to pass to the ggplot object, created by aes()}

\item{title}{A title for the plotted area}
}
\description{
Create a plot of various points in the US located in a data frame with
            an accompanying ggplot2 aesthetic (aes()).
}
